fix: generated secret name should follow release name (#10)
The generated backend auth secret and the values.yaml file
seem to be run under different contexts which will not allow
the use of `common.names.fullname` to get the name of the
current release of the chart. The secret will always grab the name
of the redhat-developer-hub chart. So instead we should use the
`.Release.Name` which will always be the release name for both.