fix: Revert setting storage class fields to `null` by default [RHIDP-5953] (#97)
* Revert setting storageClass fields to null by default

When the PVC is initially created, it gets assigned the default storage class

And it becomes impossible to upgrade the Helm Release later on because it looks like the (immutable) PVC is being patched after creation

* Bump Chart version

* Catch regression by testing an in-place upgrade of each chart from its previous version

When using the '–upgrade' flag, regression testing will only be performed if the major version of the Chart is not changed (Semver).

Otherwise, normal testing will still happen.